Ignore problems with localCheckoutDir and multi-module projects and just use AbstractMultiParentHook
